working implementation using the new gopsutil
