+++ plugins/inputs/procstat/process.go
@@ -23,6 +23,13 @@ type Process interface {
 	RlimitUsage(bool) ([]process.RlimitStat, error)
 }
 
+type PIDFinder interface {
+	PidFile(path string) ([]PID, error)
+	Pattern(pattern string) ([]PID, error)
+	Uid(user string) ([]PID, error)
+	FullPattern(path string) ([]PID, error)
+}
+
 type Proc struct {
 	hasCPUTimes bool
 	tags        map[string]string
