+++ plugins/inputs/procstat/procstat_test.go
@@ -1,3 +1,5 @@
+// +build !windows
+
 package procstat
 
 import (
